refactor + added personalization based on user info
